Refactor projects into a separate module
