+++ pt_miniscreen/pages/root/projects/__init__.py
@@ -1,0 +1,2 @@
+from .project_page import ProjectPage
+from .menu_page import ProjectsMenuPage
